pyproject.toml
[project]
name = "mcp-notes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.44",
    "mcp>=1.9.4",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
]
